refactor: rename legacy service deps
