+++ src/components/services/index.ts
@@ -1,1 +1,2 @@
-export * from "./service-category"
+export * from "./tab-content"
+export * from './_service-category';
